+++ packages/sizuku/valibot/index.ts
@@ -11,8 +11,6 @@ export const UserSchema = v.object({
   name: v.pipe(v.string(), v.minLength(1), v.maxLength(50)),
 })
 
-export type User = v.InferInput<typeof UserSchema>
-
 export const PostSchema = v.object({
   /**
    * Primary key
@@ -28,6 +26,10 @@ export const PostSchema = v.object({
   content: v.string(),
   /**
    * Foreign key referencing User.id
+   */
+  userId: v.pipe(v.string(), v.uuid()),
+})
+id
    */
   userId: v.pipe(v.string(), v.uuid()),
 })
