+++ packages/sizuku/zod/index.ts
@@ -1,5 +1,4 @@
-import { z } from 'zod/v4'
-
+import * as z from 'zod'
 export const UserSchema = z.object({
   /**
    * Primary key
